refactor(py/ai): refactored inject_instructions, moved into separate file, added tests
